v0.28.90 适配模块类型调整

+++ ibizpms-core/src/main/java/cn/ibizlab/pms/core/extensions/service/StoryExService.java
@@ -295,7 +295,7 @@ public class StoryExService extends StoryServiceImpl {
      */
     @Override
     @Transactional
-    public Story get(BigInteger key) {
+    public Story get(Long key) {
         Story et = super.get(key);
         this.getStorySpec(et);
         return et;
