+++ ibizpms-core/src/main/java/cn/ibizlab/pms/core/extensions/service/TodoExService.java
@@ -142,7 +142,7 @@ public class TodoExService extends TodoServiceImpl {
 
     @Override
     @Transactional
-    public Todo get(BigInteger key) {
+    public Todo get(Long key) {
         Todo et = getById(key);
         if(et==null){
             et=new Todo();
